// src/util/enum.ts
export enum Shape {
	SQUARE = 'square',
	ROUNDED = 'rounded'
}

export enum Size {
	SMALL = 'small',
	MEDIUM = 'medium',
	LARGE = 'large'
}

export enum Position {
	TOP = "top",
	TOP_RIGHT="top-right",
	TOP_LEFT = "top-left",
	BOTTOM ="bottom",
	BOTTOM_LEFT = "bottom-left",
	BOTTOM_RIGHT="bottom-right",
	LEFT = "left",
	LEFT_TOP = "left-top",
	LEFT_BOTTOM = "left-bottom",
	RIGHT = "right",
	RIGHT_TOP = "right-top",
	RIGHT_BOTTOM = "right-bottom",
}